# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateGuidanceStepDto:
    properties:
      description:
        type: string
      guidance_template_id:
        type: string
      step_number:
        type: integer
      title:
        type: string
    required:
    - description
    - guidance_template_id
    - step_number
    - title
    type: object
  dto.CreateGuidanceTemplateDto:
    properties:
      category:
        type: string
      description:
        type: string
      name:
        type: string
      steps:
        items:
          $ref: '#/definitions/dto.Step'
        type: array
    required:
    - description
    - name
    type: object
...
  dto.Step:
    properties:
      description:
        type: string
      id:
        type: string
      step_number:
        type: integer
      title:
        type: string
    required:
    - description
    - step_number
    - title
    type: object
  dto.UpdateAlarmDto:
    properties:
      status:
        type: string
    required:
    - status
    type: object
  dto.UpdateGuidanceTemplateDto:
    properties:
      add_steps:
        items:
          $ref: '#/definitions/dto.Step'
        type: array
      category:
        type: string
      description:
        type: string
      name:
        type: string
      remove_steps:
        items:
          type: string
        type: array
      update_steps:
        items:
          $ref: '#/definitions/dto.Step'
        type: array
    required:
    - description
    - name
    type: object
...
  errors.ErrorDetail:
    properties:
      details: {}
      message:
        type: string
      type:
        $ref: '#/definitions/errors.ErrorType'
    type: object
  errors.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/errors.ErrorDetail'
      request_id:
        type: string
      status:
        type: integer
      timestamp:
        type: string
    type: object
  errors.ErrorType:
    enum:
    - VALIDATION_ERROR
    - NOT_FOUND
    - UNAUTHORIZED
    - FORBIDDEN
    - BAD_REQUEST
    - CONFLICT
    - INTERNAL_ERROR
    - DATABASE_ERROR
    - EXTERNAL_SERVICE_ERROR
    - TIMEOUT_ERROR
    type: string
    x-enum-varnames:
    - ErrorTypeValidation
    - ErrorTypeNotFound
    - ErrorTypeUnauthorized
    - ErrorTypeForbidden
    - ErrorTypeBadRequest
    - ErrorTypeConflict
    - ErrorTypeInternal
    - ErrorTypeDatabase
    - ErrorTypeExternal
    - ErrorTypeTimeout
  models.Alarm:
    properties:
      created_at:
        type: string
      description:
        type: string
      device:
        type: string
      id:
        type: string
      premise:
        $ref: '#/definitions/models.Premise'
      premise_id:
        type: string
      severity:
        type: string
      status:
        type: string
      triggered_at:
        type: string
      type:
        type: string
    type: object
  models.GuidanceStep:
    properties:
      created_at:
        type: string
      description:
        type: string
      guidance_template:
        $ref: '#/definitions/models.GuidanceTemplate'
      guidance_template_id:
        type: string
      id:
        type: string
      step_number:
        type: integer
      title:
        type: string
    type: object
  models.GuidanceTemplate:
    properties:
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      guidance_steps:
        items:
          $ref: '#/definitions/models.GuidanceStep'
        type: array
      id:
        type: string
      name:
        type: string
    type: object
...
  models.Premise:
    properties:
      address:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      parent_premise:
        $ref: '#/definitions/models.Premise'
      parent_premise_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Smart City System (SCS) Operator API for managing premises, alarms,
    incidents, guidance templates, and guards
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: SCS Operator API
  version: "1.0"
paths:
  /alarms:
    get:
      consumes:
      - application/json
      description: Get all alarms with optional status filtering
      parameters:
      - description: Filter by alarm status (new, ignored, dispatched)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Alarm'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get alarms
      tags:
      - alarms
  /alarms/{id}:
    patch:
      consumes:
      - application/json
      description: Update an existing alarm's status or other properties
      parameters:
      - description: Alarm ID
        in: path
        name: id
        required: true
        type: string
      - description: Alarm update data
        in: body
        name: alarm
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAlarmDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Alarm'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update alarm
      tags:
      - alarms
...
  /guidance-steps:
    get:
      consumes:
      - application/json
      description: Get a list of all guidance steps
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GuidanceStep'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all guidance steps
      tags:
      - guidance-steps
    post:
      consumes:
      - application/json
      description: Create a new guidance step for a guidance template
      parameters:
      - description: Guidance step creation data
        in: body
        name: step
        required: true
        schema:
          $ref: '#/definitions/dto.CreateGuidanceStepDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.GuidanceStep'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new guidance step
      tags:
      - guidance-steps
  /guidance-steps/{id}:
    get:
      consumes:
      - application/json
      description: Get a specific guidance step by its ID
      parameters:
      - description: Guidance Step ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GuidanceStep'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get guidance step by ID
      tags:
      - guidance-steps
  /guidance-templates:
    get:
      consumes:
      - application/json
      description: Get a list of all guidance templates
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GuidanceTemplate'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all guidance templates
      tags:
      - guidance-templates
    post:
      consumes:
      - application/json
      description: Create a new guidance template with steps
      parameters:
      - description: Guidance template creation data
        in: body
        name: template
        required: true
        schema:
          $ref: '#/definitions/dto.CreateGuidanceTemplateDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.GuidanceTemplate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new guidance template
      tags:
      - guidance-templates
  /guidance-templates/{id}:
    get:
      consumes:
      - application/json
      description: Get a specific guidance template by its ID
      parameters:
      - description: Guidance Template ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GuidanceTemplate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get guidance template by ID
      tags:
      - guidance-templates
    put:
      consumes:
      - application/json
      description: Update an existing guidance template and its steps
      parameters:
      - description: Guidance Template ID
        in: path
        name: id
        required: true
        type: string
      - description: Guidance template update data
        in: body
        name: template
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateGuidanceTemplateDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GuidanceTemplate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update guidance template
      tags:
      - guidance-templates
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"